cyclictest: add CPU affinity
Add CPU affinitt option to cyclictest.

1. New option -a:

without argument to a cyclictest spreads the threads consecutive on
the available CPUs. On a quad core machine we get:

-a -t4   Thread #0 -> CPU #0
         Thread #1 -> CPU #1
         Thread #2 -> CPU #2
         Thread #3 -> CPU #3

-a -t5   Thread #0 -> CPU #0
         Thread #1 -> CPU #1
         Thread #2 -> CPU #2
         Thread #3 -> CPU #3
         Thread #4 -> CPU #0

Adding a CPU number to the -a option all threads are pinned to
the given CPU:

-a3 -t4  Thread #0 -> CPU #3
         Thread #1 -> CPU #3
         Thread #2 -> CPU #3
         Thread #3 -> CPU #3

2. extension of the -t option:

Without argument to -t cyclictest starts as many threads as CPUs are
available.

Signed-off-by: Carsten Emde <c.emde@osadl.org>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>