cyclictest: handle latest ftrace
This patch has cyclictest work a bit better with ftrace. Some new options
have been added.

When -b is used:

  -P : use the preemptoff tracer
  -I : use the irqsoff tracer
  -B : use preemptirqsoff tracer

If the above are not set, then it tries to use "events" tracer if it is
available (currently only available in the -rt kernel) and if that is not
set, it defaults to the sched-switch tracer.

-f will enabled the function tracer.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>
Cc: Darren Hart <dvhltc@us.ibm.com>
Cc: Subrata Modak <subrata@linux.vnet.ibm.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>