reworked settracer()
GeunSik Lim <leemgs1@gmail.com> found a problem with settracer()
in that it had a hardcoded limit to the number of tracers searched.
Unfortunately his patch found a bug in either fscanf or debugfs,
since it doesn't seem to return EOF at the end of input, so I reworked
the patch to read all the input and parse it using strtok.

Signed-off-by: Clark Williams <williams@redhat.com>