split function of DESTDIR into DESTDIR + prefix
It's more common that DESTDIR doesn't include /usr or /usr/local but
only the base directory for a relocated installation.  The in-tree
prefix is usually named "prefix".

"usually" here means as it is done by autoconf and as it is expected by
the Debian package helpers.

While at it also make the other path variables match this convention.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Clark Williams <williams@redhat.com>