rt-tests: Makefile: Add NUMA compile option.
This adds a NUMA compile option, and links to numa only for the tests that
need it. (Currently that is only cyclictest)

If you want to build with the NUMA feature, then define NUMA to anything.
Eg., make NUMA=1

This only adds support to the Makefile. Further patches are required
to make this work in cyclictest itself.

Signed-off-by: John Kacur <jkacur@redhat.com>