hackbench: Wait for each child explicitly by using waitpid()
Quite often it's experienced in rteval that hackbench leaves some children
as zombies during closure.  This is an attempt to keep an overview of the
status of each child separately.  It's solved by having an array with all
sender and reciever children's pids and calling waitpid() on each of these
children pairs.

Signed-off-by: David Sommerseth <davids@redhat.com>