cyclictest: Fix help for long options only
At some point in the history of cyclictest, a number of short options
were removed and changed to long only options. However the display_help
was not updated to reflect this and indicates short options that
no longer exist. Fix this. I also found a long option that wasn't listed
at all and added that.

Signed-off-by: John Kacur <jkacur@redhat.com>