Revert changes to rt-migrate-test for exit(1)
Commit 3290f8412a4562adbc05b38d8e5212090e41587e changed
exit(-1) to exit(1) for many programs in rt-tests for consistency.

rt-migrate-test needs tri-state exit statuses, because of the way
some existing bash scripts consume it's results.

This could no-doubt be improved in the future, and there are some
inconsistencies in the exit statuses in this program alone, but as they
are not urgent to fix, and the current fix makes things worse, reverting
this for now.

Signed-off-by: John Kacur <jkacur@redhat.com>
Reviewed-by: Luiz Capitulino <lcapitulino@redhat.com>