svsematest: Add duration command line argument
Many of the test programs have the --loop argument for automatic
stopping. The main problem with the --loop argument is how long is
--loop 1000?

To simplify automated tests introduce a --duration argument which
allows to set the time how long a test should run. This allows the
test suite to define the execution time and also the timeout which a
normal human can understand.

For example run the test for 10 minutes and timeout at 11 minutes:

  # timeout 11m svsematest -D 10m

Signed-off-by: Daniel Wagner <wagi@monom.org>
-Fixed some dos line endings on the man page
Signed-off-by: John Kacur <jkacur@redhat.com>