rt-tests: Allow cross compilation
$ make CROSS_COMPILE=<my-toolchain-prefix>

currently does not cause source to be built using the intended
toolchain, instead will use the host toolchain.

CC & AR are not updated with CROSS_COMPILE value since they are
predefined make variables (with default cc & ar respectively) combined
with the fact that CC & AR variables are updated in the makefile only if
not already defined due to the usage of '?='.

Signed-off-by: Afzal Mohammed <afzal.mohd.ma@gmail.com>
Signed-off-by: John Kacur <jkacur@redhat.com>