rt-tests: cyclictest: Remove support for compiling without NUMA
We announced way back in 2015 that compiling without the NUMA libs was
no-longer supported, but we left the bits in there for you to do it
anyway.

Since this is not supported, and is broken now anyway, let's remove the
cruft.

Of course running on non-NUMA machines should still work fine.

Signed-off-by: John Kacur <jkacur@redhat.com>