rt-numa: Introduce NUMA helpers
A common task is to parse the provided cpumask from the command
line. Add a helper which uses libnuma. Since we don't want to add
unnecessary dependencies for tests which don't need this helper create
a new library containing the NUMA bits.

Signed-off-by: Daniel Wagner <dwagner@suse.de>
Adding my name to the copyright in the new file src/lib/rt-numa.c
as some of the functions moved here are ones that I've written or
modified / maintained etc.
Signed-off-by: John Kacur <jkacur@redhat.com>