rt-tests: Don't compress man pages by default
This patch changes the default behavior of the Makefile to install
man pages without compression. This allows distribution build systems to
use the kind of compression they wish to use without the Makefile
getting in the way.

It is still possible to compress the man pages using the Makefile, for
example

make MAN_COMPRESSION=gzip

or

make MAN_OMPRESSION=bzip2

Signed-off-by: John Kacur <jkacur@redhat.com>