rt-tests: cyclicdeadline: Fix double mount of cgroups
If /sys/fs/cgroup exists and it is type cgroup2, cyclicdeadline mounts it a
second time as cgroup.

systemd is creating cgroup2 for logins and this can hang the machine and
not allow logins.

Fix this by:

If /sys/fs/cgroup exists, then use it for cyclicdeadline_test.

If it exists but the type is not recognized, exit with an error.
Do not simply mount it as type cgroup.

TODO: If the file doesn't exit but cgroups are supported in the kernel,
the file could be created and mounted.

Signed-off-by: John Kacur <jkacur@redhat.com>